# Conflicts:
#   client/src/main/java/client/scenes/MainCtrl.java
#   server/src/main/java/server/api/EventController.java

+++ server/src/main/java/server/api/ExpenseController.java
@@ -1,4 +1,4 @@
-package server;
+package server.api;
 
 
 import commons.Expense;
@@ -7,8 +7,10 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import server.database.ExpenseRepository;
 
+import java.util.List;
+
 @Controller
-@RequestMapping("/")
+@RequestMapping("/expense")
 public class ExpenseController {
 
     private ExpenseRepository db;
@@ -18,13 +20,23 @@ public class ExpenseController {
     }
 
     /**
+     * find all expenses
+     * @return List<Expense>
+     */
+    @GetMapping(path = { "", "/" })
+    @ResponseBody
+    public List<Expense> getAllExpenses() {
+        return db.findAll();
+    }
+
+    /**
      * Get request of the expense
      * @param expense_id id of the request
      * @return ResponseEntity<Expense> - answer of the request
      */
-    @GetMapping("/expense")
+    @GetMapping(path = {"/{id}" })
     @ResponseBody
-    public ResponseEntity<Expense> getExpenseByID(@RequestParam("id") long expense_id){
+    public ResponseEntity<Expense> getExpenseByID(@PathVariable("id") long expense_id){
         if(expense_id < 0)
             return ResponseEntity.badRequest().build();
         if(!db.existsById(expense_id))
@@ -37,9 +49,9 @@ public class ExpenseController {
     /**
      * Post request of the expense
      * @param expense - expense
-     * @return esponseEntity<Expense> - Response
+     * @return ResponseEntity<Expense> - Response
      */
-    @PostMapping()
+    @PostMapping(path = { "", "/" })
     public ResponseEntity<Expense> createExpense(@RequestBody Expense expense) {
         if (expense == null) {
             return ResponseEntity.badRequest().build();
@@ -54,18 +66,21 @@ public class ExpenseController {
      * @param expense_id - the id of the expense
      * @return ResponseEntity<Expense> - response of the method
      */
-    @PutMapping("/expense/{id}")
-    public ResponseEntity<Expense> updateExpense(@RequestBody Expense expense, @PathVariable("id") long expense_id) {
+    @PutMapping("/{id}")
+    public ResponseEntity<Expense> updateExpense(@RequestBody Expense expense,
+                                                 @PathVariable("id") long expense_id) {
         if(expense == null) {
+            return ResponseEntity.badRequest().build();
+        }
+        if(expense_id < 0) {
             return ResponseEntity.badRequest().build();
         }
         if (!db.existsById(expense_id)){
             return ResponseEntity.notFound().build();
         }
-
         Expense existingExpense = db.findById(expense_id).get();
         existingExpense.setDescription(expense.getDescription());
-        existingExpense.setValue(expense.getValue());
+        existingExpense.setAmount(expense.getAmount());
         existingExpense.setType(expense.getType());
         existingExpense.setDateSent(expense.getDateSent());
         existingExpense.setCurrency(expense.getCurrency());
@@ -78,8 +93,14 @@ public class ExpenseController {
      * @param expense_id - id of the expense
      * @return ResponseEntity<Expense> - response  of the method
      */
-    @DeleteMapping("/expense/{id}")
+    @DeleteMapping("/{id}")
     public ResponseEntity<Expense> deleteExpense(@PathVariable("id") long expense_id) {
+        if(expense_id < 0) {
+            return ResponseEntity.badRequest().build();
+        }
+        if(!db.existsById(expense_id)) {
+            return ResponseEntity.notFound().build();
+        }
         db.deleteById(expense_id);
         return ResponseEntity.noContent().build();
     }
